db changes

+++ db/migrate/20160505055722_add_price_to_orders.rb
@@ -1,5 +1,5 @@
 class AddPriceToOrders < ActiveRecord::Migration
   def change
-    add_column :orders, :price, :decimal
+    add_column :orders, :price, :integer
   end
 end
